+++ src/logic/init.ts
@@ -1,23 +1,22 @@
 import { parseProfileId } from '@superfaceai/parser';
+import { SuperJsonDocument } from '@superfaceai/sdk';
 import { basename, join as joinPath } from 'path';
 
 import {
+  BUILD_DIR,
   composeUsecaseName,
-  composeVersion,
-  EXTENSIONS,
+  GRID_DIR,
+  META_FILE,
+  NPMRC,
+  SUPERFACE_DIR,
+  TYPES_DIR,
 } from '../common/document';
 import { userError } from '../common/error';
-import { LogCallback, mkdir, mkdirQuiet, OutputStream } from '../common/io';
-import { formatShellLog } from '../common/log';
-import { ProfileSettings, ProviderSettings } from '../common/super.interfaces';
+import { mkdir, mkdirQuiet } from '../common/io';
+import { formatShellLog, LogCallback } from '../common/log';
+import { OutputStream } from '../common/output-stream';
 import * as initTemplate from '../templates/init';
 import { createProfile } from './create';
-
-export const SUPERFACE_DIR = 'superface';
-export const GRID_DIR = joinPath(SUPERFACE_DIR, 'grid');
-export const TYPES_DIR = joinPath(SUPERFACE_DIR, 'types');
-export const BUILD_DIR = joinPath(SUPERFACE_DIR, 'build');
-export const META_FILE = 'super.json';
 
 /**
  * Initializes superface at the given path.
@@ -37,8 +36,7 @@ export const META_FILE = 'super.json';
  */
 export async function initSuperface(
   appPath: string,
-  profiles: ProfileSettings,
-  providers: ProviderSettings,
+  data: SuperJsonDocument,
   options?: {
     force?: boolean;
     logCb?: LogCallback;
@@ -72,7 +70,7 @@ export async function initSuperface(
   // TODO: This will not be needed once we migrate
   // to npm repository (since it is the default)
   {
-    const npmrcPath = joinPath(appPath, '.npmrc');
+    const npmrcPath = joinPath(appPath, NPMRC);
     const created = await OutputStream.writeIfAbsent(
       npmrcPath,
       initTemplate.npmRc,
@@ -99,7 +97,7 @@ export async function initSuperface(
     const superJsonPath = joinPath(superPath, META_FILE);
     const created = await OutputStream.writeIfAbsent(
       superJsonPath,
-      () => initTemplate.superJson(profiles, providers),
+      () => initTemplate.superJson(data),
       { force: options?.force }
     );
 
@@ -150,48 +148,6 @@ export async function initSuperface(
 }
 
 /**
- * Reconstructs profile ids to correct structure for super.json
- * @param profileIds - list of profile ids
- */
-export const constructProfileSettings = (
-  profileIds: string[]
-): ProfileSettings =>
-  profileIds.reduce<ProfileSettings>((acc, profileId) => {
-    const profile = parseProfileId(profileId);
-
-    if (profile.kind === 'error') {
-      throw userError('Wrong profile Id', 1);
-    }
-
-    const { scope, name, version } = profile.value;
-    const profileName = scope ? `${scope}/${name}` : name;
-
-    acc[profileName] = {
-      version: composeVersion(version),
-      file: `./grid/${profileName}${EXTENSIONS.profile.source}`,
-    };
-
-    return acc;
-  }, {});
-
-/**
- * Reconstruct providers to correct structure for super.json
- * @param providers - list of providers
- */
-export const constructProviderSettings = (
-  providers: string[]
-): ProviderSettings =>
-  providers.reduce<ProviderSettings>((acc, provider) => {
-    acc[provider] = {
-      auth: {
-        token: 'fill-this',
-      },
-    };
-
-    return acc;
-  }, {});
-
-/**
  * Generates profiles based on profiles specified in `init` command.
  *
  * @param path - base path of folder where superface folder structure is initialized
@@ -211,12 +167,11 @@ export async function generateSpecifiedProfiles(
     }
 
     const { scope, name, version } = parsedProfile.value;
-    const profileName = scope ? `${scope}/${name}` : name;
 
     await createProfile(
       joinPath(path, GRID_DIR),
       { scope, name, version },
-      [composeUsecaseName(profileName)],
+      [composeUsecaseName(name)],
       'empty',
       { logCb }
     );